Move action setupZoomListeners to setup action file
